requirements.txt
spotipy
yt-dlp
python-dotenv
tqdm
aiohttp
pytest
rapidfuzz
